# wasm: Use cache to speed-up things