After switching to golang in guest (previously tinygo) the compile
time (InstantiateWithConfig) for wasm modules has noticeably increased.
Wazero allows using a cache to avoid compiling these modules everytime.
The cache is key is based of SHA of wasm program and cache directory uses
wazero version to ensure no surprises if IG changes wazero version. The change uses the cache across different
gadgets runs to speed-up starting a gadget. 

## Testing Done

### v0.38.0

```
$ time sudo -E ig run trace_dns:v0.38.0 --pull always --verify-image=false --timeout 1
WARN[0001] image signature verification is disabled due to using corresponding option 
WARN[0001] image signature verification is disabled due to using corresponding option 
RUNTIME.CONTAINERNAME                                            SRC                                                  DST                                                  COMM                    PID        TID QR    QTYPE                              NAME                                                             RCODE            ADDRESSES                                                         LATENCY_NS

real    0m4,274s
user    0m0,003s
sys     0m0,008s
```

### main

```
$ time sudo -E ./ig run trace_dns:main --pull always --verify-image=false --timeout 1
WARN[0001] image signature verification is disabled due to using corresponding option 
WARN[0002] image signature verification is disabled due to using corresponding option 
RUNTIME.CONTAINERNAME                                            SRC                                                  DST                                                  COMM                    PID        TID QR    QTYPE                              NAME                                                             RCODE            ADDRESSES                                                         LATENCY_NS

real    0m5,320s
user    0m0,004s
sys     0m0,007s
```

### PR

```bash
# ensure to remove cache directory
$ sudo rm -rf /var/run/ig
$ time sudo -E ./ig run trace_dns:main --pull always --verify-image=false --timeout 1
WARN[0001] image signature verification is disabled due to using corresponding option 
WARN[0002] image signature verification is disabled due to using corresponding option 
RUNTIME.CONTAINERNAME                                            SRC                                                  DST                                                  COMM                    PID        TID QR    QTYPE                              NAME                                                             RCODE            ADDRESSES                                                         LATENCY_NS

real    0m4,125s
user    0m0,004s
sys     0m0,009s

# second time things should be little more faster
$ time sudo -E ./ig run trace_dns:main --pull always --verify-image=false --timeout 1
WARN[0001] image signature verification is disabled due to using corresponding option 
WARN[0001] image signature verification is disabled due to using corresponding option 
RUNTIME.CONTAINERNAME                                            SRC                                                  DST                                                  COMM                    PID        TID QR    QTYPE                              NAME                                                             RCODE            ADDRESSES                                                         LATENCY_NS

real    0m3,448s
user    0m0,002s
sys     0m0,009s

```