@mqasimsarfraz 2025-03-24
I changed the approach of the PR, I wasn't too happy on how I was cleaning the cache especially in future we will [`Close`](https://github.com/inspektor-gadget/inspektor-gadget/pull/3778) to better handle things. Also, I realized `wazero` has file system based (I am using `tmpfs` backed location) cache which is better suited to help with gadget startup including caching it across multiple runs. 

@alban @burak-ok can you please have another look?